Module Current_docker

Integration with Docker containers.

module S : sig ... end

module Default : S.DOCKER

The default Docker engine (from the $DOCKER_HOST environment variable).

module Make (_ : S.HOST) : S.DOCKER

The docker engine running on Host.

val push_manifest : ?auth:(string * string) -> tag:string -> S.repo_id Current.t list -> S.repo_id Current.t

push_manifest images ~tag pushes a manifest containing images as tag.

parameter auth
If give, do a "docker login" using this username/password pair before pushing.

module Raw : sig ... end

Low-level API. This is useful for building custom components. The functions are similar to the ones in S.DOCKER except that each one takes the context explicitly as an argument, there are no labels, input values are no longer wrapped with Current.t, and the output is a Current.Primitive.t type. You can wrap these with let> to create your own components.