Module Default.Image

include Current_cache.S.WITH_DIGEST

type t

val digest : t -> string

digest t is a unique string for t that can be used as a primary key in a database. Two ts are considered equal if they have the same digest.

include Current_cache.S.WITH_MARSHAL with type t := t

val marshal : t -> string

Convert t to a form suitable for storage in the database.

val unmarshal : string -> t

Restore a t from a string previously produced by marshal. Raise an exception if the value cannot be read (e.g. the format has changed).

val of_hash : string -> t

val hash : t -> string

val pp : t Fmt.t